refactor: reorganise server code by data entity
Following the spirit (if not the letter) of ADR 004
