--- server/submission/index.js
@@ -1,7 +1,0 @@
-const { typeDefs } = require('./definitions')
-const resolvers = require('./resolvers')
-
-module.exports = {
-  typeDefs,
-  resolvers,
-}
