+++ server/xpub/package.json
@@ -1,10 +1,9 @@
 {
-  "name": "@elife-xpub/submission",
+  "name": "@elifesciences/xpub-server",
   "version": "1.0.0",
   "description": "",
   "license": "MIT",
-  "private": true,
-  "author": "eLife",
+  "author": "",
   "main": "index.js",
   "scripts": {},
   "keywords": [],
@@ -12,11 +11,11 @@
     "@pubsweet/component-send-email": "^0.2.3",
     "apollo-boost": "^0.1.4",
     "config": "^1.30.0",
-    "fs-extra": "^6.0.1",
+    "fs-extra": "^4.0.3",
     "graphql-tag": "^2.8.0",
-    "joi": "^13.3.0",
+    "joi": "^10.6.0",
     "lodash": "^4.17.5",
-    "request": "^2.87.0",
+    "request": "^2.85.0",
     "request-promise-native": "^1.0.5",
     "uuid": "^3.2.1",
     "xml2js": "^0.4.19"
