+++ server/xpub/schema/xpub.graphqls
